straus_lookup_table in new files
